requirements.txt
Django>=3.0.4,<3.1.0
djangorestframework>=3.11.0,<3.12.0
selenium>=3.14.0,<3.20.0

flake8>=3.6.0,<3.7.0
